Implement cron job for counter updates every 10 minutes and streamline counter retrieval logic
